pyproject.toml
[tool.poetry]
name = "python-final-diplom"
version = "0.1.0"
description = ""
authors = ["Andrey Shishov <a.5920311@yandex.ru>"]
readme = "README.md"
packages = [{include = "orders"}]

[tool.poetry.dependencies]
python = "^3.10"
pytz = "2018.5"
environ = "^1.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pytest-django = "^4.5.2"
pytest = "^7.2.2"
model-bakery = "^1.10.1"
pyyaml = "^6.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "orders.settings"
pythonpath = ". orders"